requirements.txt
h5py==3.12.1
keras==3.6.0
matplotlib==3.10.3
numpy==2.3.1
Pillow==11.2.1
streamlit==1.44.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
